Fix TS issue with a verbose solution
