+++ tests/common.spec.ts
@@ -1,5 +1,5 @@
-import { beforeEach, expect, test } from "@jest/globals";
-import { constructPaginationString } from "../src/common";
+import { expect, test } from "@jest/globals";
+import { constructPaginationString } from "../src/utils/pagination";
 
 test("constructs url without query params", async () => {
   const url = "https://google.com";
@@ -37,7 +37,7 @@ test("constructs url with query params (limit and next)", async () => {
     url,
     paginationParameters: {
       limit: 10,
-      next: 20
+      next: 20,
     },
   });
   expect(result).toEqual(url + "?limit=10&until=20");
@@ -49,11 +49,11 @@ test("constructs url with query params (limit and previous)", async () => {
     url,
     paginationParameters: {
       limit: 4,
-      previous: 5
-    }
-  })
-  expect(result).toEqual(url + "?limit=4&until=5")
-})
+      previous: 5,
+    },
+  });
+  expect(result).toEqual(url + "?limit=4&until=5");
+});
 
 test("constructs url with query params (next only)", async () => {
   const url = "https://google.com";
@@ -61,10 +61,10 @@ test("constructs url with query params (next only)", async () => {
     url,
     paginationParameters: {
       next: 5,
-    }
-  })
-  expect(result).toBe(url + "?until=5")
-})
+    },
+  });
+  expect(result).toBe(url + "?until=5");
+});
 
 test("constructs url with query params (previous only)", async () => {
   const url = "https://google.com";
@@ -72,8 +72,7 @@ test("constructs url with query params (previous only)", async () => {
     url,
     paginationParameters: {
       previous: 2,
-    }
-  })
-  expect(result).toBe(url + "?until=2")
-
-})
+    },
+  });
+  expect(result).toBe(url + "?until=2");
+});
